fw: v6.79.1 -- points_per_sec did its arithmetic in a container two sizes too small
pointsPerSec() computed (delta_points * 1000) / elapsed in 32-bit. The trap is
that its averaging window is not 1 s: nothing recomputes the counter on a
timer, so the window is however long it has been since something last READ it.
Close the WebUI and stop polling /api/state and that window grows without
bound -- at ~45 kpps a gap over ~95 s pushes delta * 1000 past 2^32 and the
field wraps.

Measured on this board: first reads after 100 s and 140 s of silence returned
2457 and 14694 pps against a true 45355, matching the wrapped arithmetic to
within 0.1%, while 30 s and 60 s gaps read clean.

Not cosmetic. optimizeGalvo derives the camera shutter from this field, and a
4.8 kpps reading made it compute a 251.6 ms frame period for a pattern that
actually takes 26.8 ms -- which then derived a 500 ms exposure and nearly
saturated the background. It also quietly undermined the "points_per_sec vs
projection.kpps is the one-line health check" rule, which was unreliable
precisely when nobody was watching.

Both counters now use the (uint64_t) ... * 1000ULL idiom autotuneTrial()
already uses a few hundred lines up. fps() shares the shape but not the risk
(~30 fps needs 40 hours to overflow); changed anyway, because there is no good
reason for two counters sitting next to each other to differ in correctness.

The tool is separately immunized by double-reading points_per_sec, so a device
still on older firmware is unaffected.

+++ src/output/galvo_out.cpp
@@ -1085,7 +1085,18 @@ bool pushFrame(const LaserPoint* pts, size_t count) {
 uint32_t pointsPerSec() {
     uint32_t now = millis();
     if (now - s_last_pps_t >= 1000) {
-        s_pps_cached  = (s_points_total - s_points_last) * 1000 / (now - s_last_pps_t);
+        // 64-bit multiply: the averaging window is NOT 1 s, it is however long it has
+        // been since something last READ this counter -- nothing recomputes it on a
+        // timer. With no poller (WebUI closed, no /api/state traffic) that window grows
+        // without bound, and at ~45 kpps a gap over ~95 s makes delta * 1000 exceed
+        // 2^32 in 32-bit arithmetic, so the field wraps to a garbage value. Measured on
+        // this board: a first read after several minutes idle returned 2457 and 14694
+        // pps against a true 45355, matching the wrapped arithmetic to within 0.1%.
+        // That is not cosmetic -- optimizeGalvo derives the camera shutter from this
+        // field, and a 4.8 kpps reading made it compute a 251.6 ms frame period for a
+        // pattern that actually takes 26.8 ms.
+        s_pps_cached  = (uint32_t)(((uint64_t)(s_points_total - s_points_last) * 1000ULL)
+                                   / (uint64_t)(now - s_last_pps_t));
         s_points_last = s_points_total;
         s_last_pps_t  = now;
         gState.points_per_sec = s_pps_cached;
@@ -1098,7 +1109,12 @@ uint32_t pointsPerSec() {
 uint32_t fps() {
     uint32_t now = millis();
     if (now - s_last_fps_t >= 1000) {
-        s_fps_cached  = (s_frames_total - s_frames_last) * 1000 / (now - s_last_fps_t);
+        // Same unbounded window as pointsPerSec() above, so the same 64-bit multiply.
+        // Frames accumulate ~1500x slower than points, which puts the overflow past 40
+        // hours of silence rather than 95 seconds -- latent rather than reachable, but
+        // there is no reason for the two counters to differ in correctness.
+        s_fps_cached  = (uint32_t)(((uint64_t)(s_frames_total - s_frames_last) * 1000ULL)
+                                   / (uint64_t)(now - s_last_fps_t));
         s_frames_last = s_frames_total;
         s_last_fps_t  = now;
         gState.fps = s_fps_cached;
